requirements.txt
streamlit==1.37.1
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
pypdf==4.3.1
chromadb==0.5.5
tiktoken==0.7.0
python-dotenv==1.0.1
pydantic==2.8.2
